+++ cases/testcases/boxed_math/red.clj
@@ -1,0 +1,4 @@
+(ns testcases.boxed-math.red)
+
+(defn sum-squares [a b]
+  (+ (* a a) (* b b)))